chatbot/pyproject.toml
[project]
name = "techathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cryptography==46.0.5",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "fastapi==0.129.0",
    "filetype==1.2.0",
    "frozenlist==1.8.0",
    "google-api-core==2.30.0",
    "google-api-python-client==2.190.0",
    "google-auth==2.48.0",
    "google-auth-httplib2==0.3.0",
    "google-cloud-core==2.5.0",
    "google-cloud-modelarmor==0.4.0",
    "google-genai==1.64.0",
    "googleapis-common-protos==1.72.0",
    "googlemaps==4.10.0",
    "greenlet==3.3.1",
    "groq==0.37.1",
    "grpcio==1.78.1",
    "grpcio-status==1.78.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.2",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.14",
    "langchain-google-community==3.0.5",
    "langchain-google-genai==4.2.1",
    "langchain-groq==1.1.2",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.9",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.8",
    "langsmith==0.7.5",
    "marshmallow==3.26.2",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.2",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pillow==12.1.1",
    "propcache==0.4.1",
    "proto-plus==1.27.1",
    "protobuf==6.33.5",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pyparsing==3.3.2",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "reportlab==4.4.10",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "starlette==0.52.1",
    "tenacity==9.1.4",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "uvicorn==0.41.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
